Remove legacy files and add new address_info module.
Replaced outdated `api_spec.py` and related legacy documentation with a new `address_info` module. This module enables retrieving information about address objects in SCM with enhanced filtering options, improving modularity and usability.
