+++ .ansible-lint
@@ -30,6 +30,9 @@ skip_list:
   - role-name
   - internal-error
   - galaxy
+  - '204'
+  - 'ignore-errors'
+  - 'var-naming[no-role-prefix]'
 
 # Set verbosity level
 verbosity: 1
@@ -49,6 +52,7 @@ warn_list:
   - no-changed-when
   - no-handler
   - var-naming[no-role-prefix]
+  - experimental
 
 # Custom settings (adjust as needed)
 parseable: true
