+++ .devcontainer/Dockerfile
@@ -1,0 +1,23 @@
+# Note: You can use any Debian/Ubuntu based image you want.
+FROM mcr.microsoft.com/devcontainers/python:3.12-bullseye
+
+COPY library-scripts/*.sh /tmp/library-scripts/
+COPY requirements.txt /tmp/requirements.txt
+
+# Install needed packages and setup non-root user. Use a separate RUN statement to add your own dependencies.
+ARG USERNAME=vscode
+ARG USER_UID=1000
+ARG USER_GID=$USER_UID
+RUN apt-get update && export DEBIAN_FRONTEND=noninteractive \
+    && bash /tmp/library-scripts/common-debian.sh "${INSTALL_ZSH}" "${USERNAME}" "${USER_UID}" "${USER_GID}" "${UPGRADE_PACKAGES}" "true" "true" \
+    && apt-get install -y libssl-dev libffi-dev python3-dev python3-pip vim \
+    && apt-get clean -y && rm -rf /var/lib/apt/lists/*
+
+# Install Ansible + Dependencies
+RUN python3 -m pip install -r /tmp/requirements.txt
+
+CMD [ "sleep", "infinity" ]
+
+# [Optional] Uncomment this section to install additional OS packages.
+# RUN apt-get update && export DEBIAN_FRONTEND=noninteractive \
+#     && apt-get -y install --no-install-recommends <your-package-list-here>
