+++ .devcontainer/library-scripts/README.md
@@ -1,0 +1,7 @@
+# Warning: Folder contents may be replaced
+
+The contents of this folder will be automatically replaced with a file of the same name in the
+repository's [script-library folder](https://github.com/microsoft/vscode-dev-containers/tree/main/script-library)
+whenever the repository is packaged.
+
+To retain your edits, move the file to a different location. You may also delete the files if they are not needed.
