+++ .github/workflows/ansible-lint.yml
@@ -1,0 +1,15 @@
+name: Ansible Lint
+
+on:
+  push:
+    branches: [ main ]
+  pull_request:
+    branches: [ main ]
+
+jobs:
+  ansible_lint:
+    name: Verify Ansible Lint
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+      - uses: ansible/ansible-lint@v25.1.1
