+++ .github/workflows/ci.yml
@@ -2,9 +2,9 @@ name: Ansible Collection CI
 
 on:
   push:
-    branches: [main]
+    branches: [ main ]
   pull_request:
-    branches: [main]
+    branches: [ main ]
 
 jobs:
   lint:
@@ -31,127 +31,3 @@ jobs:
           # Only run yamllint for now, skip ansible-lint until module resolution is fixed
           echo "Skipping ansible-lint due to module resolution issues"
           yamllint .
-
-  build-and-test:
-    runs-on: ubuntu-latest
-    needs: lint
-    strategy:
-      matrix:
-        ansible-version: ['2.15', '2.16', '2.17', 'latest']
-    steps:
-      - uses: actions/checkout@v3
-      - name: Set up Python
-        uses: actions/setup-python@v4
-        with:
-          python-version: '3.11'
-      - name: Install dependencies
-        run: |
-          python -m pip install --upgrade pip
-          if [ "${{ matrix.ansible-version }}" == "latest" ]; then
-            pip install ansible-core
-          else
-            pip install ansible-core~=${{ matrix.ansible-version }}.0
-          fi
-          pip install poetry
-          poetry install
-      - name: Build collection
-        run: ansible-galaxy collection build pan_scm_ansible
-      - name: Install collection and dependencies
-        run: |
-          ansible-galaxy collection install *.tar.gz -p ./collections --force
-          cd ./collections/ansible_collections/cdot65/scm
-          pip install -r ../../../../pan_scm_ansible/requirements.txt
-          # Fix for module imports in test environment
-          touch __init__.py
-          cd ../../../../
-      - name: Create sanity ignore directory
-        run: |
-          mkdir -p ./collections/ansible_collections/cdot65/scm/tests/sanity
-          cp pan_scm_ansible/tests/sanity/ignore-2.18.txt ./collections/ansible_collections/cdot65/scm/tests/sanity/
-      - name: Create setup script for test containers
-        run: |
-          cat > setup_test_container.sh << 'EOF'
-          #!/bin/bash
-          # Script to prepare test container with required dependencies
-          pip install pan-scm-sdk>=0.3.22
-          EOF
-          chmod +x setup_test_container.sh
-      - name: Run sanity tests with ignore file
-        run: |
-          cd ./collections/ansible_collections/cdot65/scm
-          export PYTHONPATH=$(pwd)
-          ansible-test sanity --docker -v --docker-setup-file ../../../../setup_test_container.sh
-      - name: Run unit tests
-        run: |
-          cd ./collections/ansible_collections/cdot65/scm
-          ansible-test units --docker -v --docker-setup-file ../../../../setup_test_container.sh
-  
-  integration-tests:
-    runs-on: ubuntu-latest
-    needs: build-and-test
-    if: github.event_name == 'push' && github.ref == 'refs/heads/main'
-    steps:
-      - uses: actions/checkout@v3
-      - name: Set up Python
-        uses: actions/setup-python@v4
-        with:
-          python-version: '3.11'
-      - name: Install dependencies
-        run: |
-          python -m pip install --upgrade pip
-          pip install ansible-core
-          pip install poetry
-          poetry install
-      - name: Build collection
-        run: ansible-galaxy collection build pan_scm_ansible
-      - name: Install collection and dependencies
-        run: |
-          ansible-galaxy collection install *.tar.gz -p ./collections --force
-          cd ./collections/ansible_collections/cdot65/scm
-          pip install -r ../../../../pan_scm_ansible/requirements.txt
-          # Fix for module imports in test environment
-          touch __init__.py
-          cd ../../../../
-      - name: Create sanity ignore directory
-        run: |
-          mkdir -p ./collections/ansible_collections/cdot65/scm/tests/sanity
-          cp pan_scm_ansible/tests/sanity/ignore-2.18.txt ./collections/ansible_collections/cdot65/scm/tests/sanity/
-      - name: Create setup script for test containers
-        run: |
-          cat > setup_test_container.sh << 'EOF'
-          #!/bin/bash
-          # Script to prepare test container with required dependencies
-          pip install pan-scm-sdk>=0.3.22
-          EOF
-          chmod +x setup_test_container.sh
-      - name: Run integration tests
-        run: |
-          cd ./collections/ansible_collections/cdot65/scm
-          export PYTHONPATH=$(pwd)
-          ansible-test integration --docker -v --docker-setup-file ../../../../setup_test_container.sh
-        env:
-          PAN_SCM_USERNAME: ${{ secrets.PAN_SCM_USERNAME }}
-          PAN_SCM_PASSWORD: ${{ secrets.PAN_SCM_PASSWORD }}
-          PAN_SCM_TENANT: ${{ secrets.PAN_SCM_TENANT }}
-
-  documentation:
-    runs-on: ubuntu-latest
-    needs: lint
-    steps:
-      - uses: actions/checkout@v3
-      - name: Set up Python
-        uses: actions/setup-python@v4
-        with:
-          python-version: '3.11'
-      - name: Install dependencies
-        run: |
-          python -m pip install --upgrade pip
-          pip install mkdocs-material mkdocstrings-python mkdocs-autorefs termynal
-      - name: Build documentation
-        run: mkdocs build --verbose
-      - name: Deploy documentation
-        if: github.event_name == 'push' && github.ref == 'refs/heads/main'
-        uses: peaceiris/actions-gh-pages@v3
-        with:
-          github_token: ${{ secrets.GITHUB_TOKEN }}
-          publish_dir: ./site
