+++ .github/workflows/yamllint.yml
@@ -1,0 +1,19 @@
+name: YAML Lint
+
+on:
+  push:
+    branches: [ main ]
+  pull_request:
+    branches: [ main ]
+
+jobs:
+  yamllint:
+    name: Verify yamllint
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+      - uses: actions/setup-python@v5
+        with:
+          python-version: "3.x"
+      - name: Run yaml-lint
+        uses: ibiqlik/action-yamllint@v3
