+++ ISSUES.md
@@ -2,25 +2,41 @@
 
 ## Reporting
 
-Use the GitHub [issue tracker](https://github.com/cdot65/pan-scm-ansible-collection/issues) for filing bugs. In order to save time and help us respond to issues quickly, make sure to fill out as much of the issue template as possible. Version information and an accurate reproduction scenario are critical to helping us identify the problem.
+Use the GitHub [issue tracker](https://github.com/cdot65/pan-scm-ansible/issues) for filing bugs. In order to save time
+and help us respond to issues quickly, make sure to fill out as much of the issue template as possible. Version
+information and an accurate reproduction scenario are critical to helping us identify the problem.
 
-Please don't use the issue tracker as a way to ask how to do something. Instead, use the [GitHub Discussion Forum](https://github.com/cdot65/pan-scm-ansible-collection/discussion).
+Please don't use the issue tracker as a way to ask how to do something. Instead, use
+the [GitHub Discussion Forum](https://github.com/cdot65/pan-scm-ansible/discussion).
 
-Before opening a new issue, please use the issue search feature to see if what you're experiencing has already been reported. If you have any extra detail to provide, please comment. Otherwise, rather than posting a "me too" comment, consider giving it a ["thumbs up"](https://github.blog/2016-03-10-add-reactions-to-pull-requests-issues-and-comments) to give us an indication of the severity of the problem.
+Before opening a new issue, please use the issue search feature to see if what you're experiencing has already been
+reported. If you have any extra detail to provide, please comment. Otherwise, rather than posting a "me too" comment,
+consider giving it a ["thumbs up"](https://github.blog/2016-03-10-add-reactions-to-pull-requests-issues-and-comments) to
+give us an indication of the severity of the problem.
 
 ## How Issues Are Resolved
 
-We triage our issues into high, medium, and low, and tag them with the relevant component. We typically focus on higher priority issues first. There aren't hard and fast rules for determining the severity of an issue, but generally high priority issues have an increased likelihood of breaking existing functionality and negatively impacting a large number of users.
+We triage our issues into high, medium, and low, and tag them with the relevant component. We typically focus on higher
+priority issues first. There aren't hard and fast rules for determining the severity of an issue, but generally high
+priority issues have an increased likelihood of breaking existing functionality and negatively impacting a large number
+of users.
 
 If your issue isn't considered high priority, please be patient as it may take some time to get to it.
 
 ### Issue States
 
-- **`state:needs_triage`**: This issue has not been looked at by a person yet and still needs to be triaged. This is the initial state for all new issues/pull requests.
-- **`state:needs_info`**: The issue needs more information, such as more debug output or more specifics about the system, such as version information. Any detail that is currently preventing this issue from moving forward. This should be considered a blocked state.
-- **`state:needs_review`**: The issue/pull request needs to be reviewed by other maintainers and contributors. This is usually used when there is a question to another maintainer or when a person is less familiar with an area of the code base.
-- **`state:needs_revision`**: More commonly used on pull requests, this state represents that there are changes being waited on.
-- **`state:in_progress`**: The issue is actively being worked on and you should be in contact with whoever is assigned if you are also working on or plan to work on a similar issue.
+- **`state:needs_triage`**: This issue has not been looked at by a person yet and still needs to be triaged. This is the
+  initial state for all new issues/pull requests.
+- **`state:needs_info`**: The issue needs more information, such as more debug output or more specifics about the
+  system, such as version information. Any detail that is currently preventing this issue from moving forward. This
+  should be considered a blocked state.
+- **`state:needs_review`**: The issue/pull request needs to be reviewed by other maintainers and contributors. This is
+  usually used when there is a question to another maintainer or when a person is less familiar with an area of the code
+  base.
+- **`state:needs_revision`**: More commonly used on pull requests, this state represents that there are changes being
+  waited on.
+- **`state:in_progress`**: The issue is actively being worked on and you should be in contact with whoever is assigned
+  if you are also working on or plan to work on a similar issue.
 - **`state:in_testing`**: The issue or pull request is currently being tested.
 
 #### For Issue Submitters
@@ -31,7 +47,9 @@ Please use the issue template and provide as much information as possible:
 - Component
 - Summary
 
-If any of those items are missing, your issue will still get the `needs_triage` label but may end up being responded to slower than issues that have the complete set of information. The `state:needs_triage` label will remain on your issue until a person has looked at it.
+If any of those items are missing, your issue will still get the `needs_triage` label but may end up being responded to
+slower than issues that have the complete set of information. The `state:needs_triage` label will remain on your issue
+until a person has looked at it.
 
 #### For Pull Request Submitters
 
@@ -41,6 +59,7 @@ Please use the pull request template and provide as much information as possible
 - Component
 - Summary
 
-If any of those items are missing, your pull request will still get the `needs_triage` label but may end up being responded to slower than other pull requests that have a complete set of information.
+If any of those items are missing, your pull request will still get the `needs_triage` label but may end up being
+responded to slower than other pull requests that have a complete set of information.
 
 The `state:needs_triage` label will remain on your pull request until a person has looked at it.
