+++ meta/runtime.yml
@@ -1,0 +1,2 @@
+---
+requires_ansible: ">=2.18.0"
