--- pan_scm_ansible/meta/runtime.yml
@@ -1,2 +1,0 @@
----
-requires_ansible: ">=2.14.0"
