+++ plugins/modules/address_info.py
@@ -1,0 +1,339 @@
+# -*- coding: utf-8 -*-
+
+# This code is part of Ansible, but is an independent component.
+# This particular file snippet, and this file snippet only, is Apache2.0 licensed.
+# Modules you write using this snippet, which is embedded dynamically by Ansible
+# still belong to the author of the module, and may assign their own license
+# to the complete work.
+#
+# Copyright (c) 2024 Calvin Remsburg (@cdot65)
+# All rights reserved.
+
+"""
+Ansible module for gathering information about address objects in SCM.
+
+This module provides functionality to retrieve information about address objects
+in the SCM (Strata Cloud Manager) system with various filtering options.
+"""
+
+from __future__ import absolute_import, division, print_function
+
+__metaclass__ = type
+
+from ansible.module_utils.basic import AnsibleModule
+from ansible.module_utils.common.text.converters import to_text
+from ansible_collections.cdot65.scm.plugins.module_utils.authenticate import get_scm_client
+from ansible_collections.cdot65.scm.plugins.module_utils.serialize_response import serialize_response
+from scm.exceptions import InvalidObjectError, MissingQueryParameterError, ObjectNotPresentError
+
+DOCUMENTATION = r"""
+---
+module: address_info
+
+short_description: Gather information about address objects in SCM.
+
+version_added: "0.1.0"
+
+description:
+    - Gather information about address objects within Strata Cloud Manager (SCM).
+    - Supports retrieving a specific address by name or listing addresses with various filters.
+    - Provides additional client-side filtering capabilities for exact matches and exclusions.
+    - Returns detailed information about each address object.
+
+options:
+    name:
+        description: The name of a specific address object to retrieve.
+        required: false
+        type: str
+    folder:
+        description: Filter addresses by folder container.
+        required: false
+        type: str
+    snippet:
+        description: Filter addresses by snippet container.
+        required: false
+        type: str
+    device:
+        description: Filter addresses by device container.
+        required: false
+        type: str
+    exact_match:
+        description: When True, only return objects defined exactly in the specified container.
+        required: false
+        type: bool
+        default: false
+    exclude_folders:
+        description: List of folder names to exclude from results.
+        required: false
+        type: list
+        elements: str
+    exclude_snippets:
+        description: List of snippet values to exclude from results.
+        required: false
+        type: list
+        elements: str
+    exclude_devices:
+        description: List of device values to exclude from results.
+        required: false
+        type: list
+        elements: str
+    types:
+        description: Filter by address types.
+        required: false
+        type: list
+        elements: str
+        choices: ["netmask", "range", "wildcard", "fqdn"]
+    values:
+        description: Filter by address values.
+        required: false
+        type: list
+        elements: str
+    tags:
+        description: Filter by tags.
+        required: false
+        type: list
+        elements: str
+    provider:
+        description: Authentication credentials.
+        required: true
+        type: dict
+        suboptions:
+            client_id:
+                description: Client ID for authentication.
+                required: true
+                type: str
+            client_secret:
+                description: Client secret for authentication.
+                required: true
+                type: str
+            tsg_id:
+                description: Tenant Service Group ID.
+                required: true
+                type: str
+            log_level:
+                description: Log level for the SDK.
+                required: false
+                type: str
+                default: "INFO"
+
+author:
+    - Calvin Remsburg (@cdot65)
+"""
+
+EXAMPLES = r"""
+---
+- name: Gather Address Information in Strata Cloud Manager
+  hosts: localhost
+  gather_facts: false
+  vars_files:
+    - vault.yaml
+  vars:
+    provider:
+      client_id: "{{ client_id }}"
+      client_secret: "{{ client_secret }}"
+      tsg_id: "{{ tsg_id }}"
+      log_level: "INFO"
+  tasks:
+
+    - name: Get information about a specific address
+      cdot65.scm.address_info:
+        provider: "{{ provider }}"
+        name: "web-server"
+        folder: "Texas"
+      register: address_info
+
+    - name: List all address objects in a folder
+      cdot65.scm.address_info:
+        provider: "{{ provider }}"
+        folder: "Texas"
+      register: all_addresses
+
+    - name: List only FQDN address objects
+      cdot65.scm.address_info:
+        provider: "{{ provider }}"
+        folder: "Texas"
+        types: ["fqdn"]
+      register: fqdn_addresses
+
+    - name: List addresses with specific tags
+      cdot65.scm.address_info:
+        provider: "{{ provider }}"
+        folder: "Texas"
+        tags: ["Production", "Web"]
+      register: tagged_addresses
+
+    - name: List addresses with exact match and exclusions
+      cdot65.scm.address_info:
+        provider: "{{ provider }}"
+        folder: "Texas"
+        exact_match: true
+        exclude_folders: ["All"]
+        exclude_snippets: ["default"]
+      register: filtered_addresses
+"""
+
+RETURN = r"""
+addresses:
+    description: List of address objects matching the filter criteria (returned when name is not specified).
+    returned: success, when name is not specified
+    type: list
+    elements: dict
+    sample:
+      - id: "123e4567-e89b-12d3-a456-426655440000"
+        name: "web-server"
+        description: "Web server address"
+        ip_netmask: "192.168.1.100/32"
+        folder: "Texas"
+        tag: ["Web", "Production"]
+      - id: "234e5678-e89b-12d3-a456-426655440001"
+        name: "app-server"
+        description: "Application server address"
+        ip_netmask: "192.168.1.101/32"
+        folder: "Texas"
+        tag: ["App", "Production"]
+address:
+    description: Information about the requested address (returned when name is specified).
+    returned: success, when name is specified
+    type: dict
+    sample:
+        id: "123e4567-e89b-12d3-a456-426655440000"
+        name: "web-server"
+        description: "Web server address"
+        ip_netmask: "192.168.1.100/32"
+        folder: "Texas"
+        tag: ["Web", "Production"]
+"""
+
+
+def build_filter_params(module_params):
+    """
+    Build filter parameters dictionary from module parameters.
+
+    Args:
+        module_params (dict): Dictionary of module parameters
+
+    Returns:
+        dict: Filtered dictionary containing only relevant filter parameters
+    """
+    # Container params
+    container_params = {}
+    for container in ["folder", "snippet", "device"]:
+        if module_params.get(container) is not None:
+            container_params[container] = module_params[container]
+
+    # Filter params
+    filter_params = {}
+    for filter_param in ["exact_match", "exclude_folders", "exclude_snippets", "exclude_devices"]:
+        if module_params.get(filter_param) is not None:
+            filter_params[filter_param] = module_params[filter_param]
+
+    # Convert "tags" parameter to "tags" for SDK
+    if module_params.get("tags") is not None:
+        filter_params["tags"] = module_params["tags"]
+
+    # Add other filter parameters
+    for param in ["types", "values"]:
+        if module_params.get(param) is not None:
+            filter_params[param] = module_params[param]
+
+    return container_params, filter_params
+
+
+def main():
+    """
+    Main execution path for the address_info module.
+
+    This module provides functionality to gather information about address objects
+    in the SCM (Strata Cloud Manager) system with various filtering options.
+
+    :return: Ansible module exit data
+    :rtype: dict
+    """
+    module = AnsibleModule(
+        argument_spec=dict(
+            name=dict(type="str", required=False),
+            folder=dict(type="str", required=False),
+            snippet=dict(type="str", required=False),
+            device=dict(type="str", required=False),
+            exact_match=dict(type="bool", required=False, default=False),
+            exclude_folders=dict(type="list", elements="str", required=False),
+            exclude_snippets=dict(type="list", elements="str", required=False),
+            exclude_devices=dict(type="list", elements="str", required=False),
+            types=dict(type="list", elements="str", required=False),
+            values=dict(type="list", elements="str", required=False),
+            tags=dict(type="list", elements="str", required=False),
+            provider=dict(
+                type="dict",
+                required=True,
+                options=dict(
+                    client_id=dict(type="str", required=True),
+                    client_secret=dict(type="str", required=True, no_log=True),
+                    tsg_id=dict(type="str", required=True),
+                    log_level=dict(type="str", required=False, default="INFO"),
+                ),
+            ),
+        ),
+        supports_check_mode=True,
+        mutually_exclusive=[
+            ["folder", "snippet", "device"]
+        ],
+        required_one_of=[
+            ["folder", "snippet", "device"]
+        ],
+    )
+
+    result = {}
+
+    try:
+        client = get_scm_client(module)
+        
+        # Check if we're fetching a specific address by name
+        if module.params.get("name"):
+            name = module.params["name"]
+            container_params = {}
+            
+            # Get the container param
+            for container in ["folder", "snippet", "device"]:
+                if module.params.get(container):
+                    container_params[container] = module.params[container]
+            
+            try:
+                # Fetch a specific address
+                address = client.address.fetch(
+                    name=name,
+                    **container_params
+                )
+                
+                # Serialize response for Ansible output
+                result["address"] = serialize_response(address)
+                
+            except ObjectNotPresentError:
+                module.fail_json(
+                    msg=f"Address with name '{name}' not found in {list(container_params.keys())[0]} '{list(container_params.values())[0]}'"
+                )
+            except (MissingQueryParameterError, InvalidObjectError) as e:
+                module.fail_json(msg=str(e))
+                
+        else:
+            # List addresses with filtering
+            container_params, filter_params = build_filter_params(module.params)
+            
+            try:
+                addresses = client.address.list(**container_params, **filter_params)
+                
+                # Serialize response for Ansible output
+                result["addresses"] = [serialize_response(addr) for addr in addresses]
+                
+            except MissingQueryParameterError as e:
+                module.fail_json(msg=f"Missing required parameter: {str(e)}")
+            except InvalidObjectError as e:
+                module.fail_json(msg=f"Invalid filter parameters: {str(e)}")
+        
+        module.exit_json(**result)
+        
+    except Exception as e:
+        module.fail_json(msg=to_text(e))
+
+
+if __name__ == "__main__":
+    main()
