+++ plugins/modules/application.py
@@ -13,7 +13,7 @@
 Ansible module for managing application objects in SCM.
 
 This module provides functionality to create, update, and delete application objects
-in the SCM (Security Control Manager) system. It handles various application attributes
+in the SCM (Strata Cloud Manager) system. It handles various application attributes
 and supports check mode operations.
 """
 
@@ -23,7 +23,7 @@ __metaclass__ = type
 
 from ansible.module_utils._text import to_text
 from ansible.module_utils.basic import AnsibleModule
-from ansible_collections.cdot65.scm.plugins.module_utils.api_spec import ScmSpec  # noqa: F401
+from ansible_collections.cdot65.scm.plugins.module_utils.api_spec import ApplicationSpec  # noqa: F401
 from ansible_collections.cdot65.scm.plugins.module_utils.authenticate import (  # noqa: F401
     get_scm_client,
 )
@@ -275,7 +275,7 @@ def main():
     Main execution path for the application object module.
     """
     module = AnsibleModule(
-        argument_spec=ScmSpec.application_spec(),
+        argument_spec=ApplicationSpec.spec(),
         supports_check_mode=True,
         required_if=[
             ("state", "present", ["category", "subcategory", "technology", "risk"]),
@@ -312,8 +312,8 @@ def main():
                 need_update = False
                 for key, value in application_data.items():
                     if (
-                        hasattr(existing_application, key)
-                        and getattr(existing_application, key) != value
+                            hasattr(existing_application, key)
+                            and getattr(existing_application, key) != value
                     ):
                         need_update = True
                         break
