+++ plugins/modules/application_group.py
@@ -10,10 +10,10 @@
 # All rights reserved.
 
 """
-Ansible module for managing service group objects in SCM.
+Ansible module for managing application group objects in SCM.
 
-This module provides functionality to create, update, and delete service group objects
-in the SCM (Security Control Manager) system. It handles service group members
+This module provides functionality to create, update, and delete application group objects
+in the SCM (Strata Cloud Manager) system. It handles application group membership
 and supports check mode operations.
 """
 
@@ -23,7 +23,8 @@ __metaclass__ = type
 
 from ansible.module_utils._text import to_text
 from ansible.module_utils.basic import AnsibleModule
-from ansible_collections.cdot65.scm.plugins.module_utils.api_spec import ScmSpec  # noqa: F401
+from ansible_collections.cdot65.scm.plugins.module_utils.api_spec.application_group import \
+    ApplicationGroupSpec  # noqa: F401
 from ansible_collections.cdot65.scm.plugins.module_utils.authenticate import (  # noqa: F401
     get_scm_client,
 )
@@ -32,37 +33,35 @@ from ansible_collections.cdot65.scm.plugins.module_utils.serialize_response impo
 )
 from pydantic import ValidationError
 
-from scm.config.objects.service_group import ServiceGroup
+from scm.config.objects.application_group import ApplicationGroup
 from scm.exceptions import NotFoundError
-from scm.models.objects.service_group import ServiceGroupCreateModel, ServiceGroupUpdateModel
+from scm.models.objects.application_group import (
+    ApplicationGroupCreateModel,
+    ApplicationGroupUpdateModel,
+)
 
 DOCUMENTATION = r"""
 ---
-module: service_group
+module: application_group
 
-short_description: Manage service group objects in SCM.
+short_description: Manage application group objects in SCM.
 
 version_added: "0.1.0"
 
 description:
-    - Manage service group objects within Strata Cloud Manager (SCM).
-    - Supports creation, modification, and deletion of service group objects.
-    - Ensures proper validation of service group members.
+    - Manage application group objects within Strata Cloud Manager (SCM).
+    - Supports creation, modification, and deletion of application groups.
+    - Ensures proper validation of group membership.
     - Ensures that exactly one of 'folder', 'snippet', or 'device' is provided.
 
 options:
     name:
-        description: The name of the service group.
+        description: The name of the application group.
         required: true
         type: str
     members:
-        description: List of service objects that are members of this group.
+        description: List of application / group / filter names.
         required: true
-        type: list
-        elements: str
-    tag:
-        description: List of tags associated with the service group.
-        required: false
         type: list
         elements: str
     folder:
@@ -101,7 +100,7 @@ options:
                 type: str
                 default: "INFO"
     state:
-        description: Desired state of the service group object.
+        description: Desired state of the application group object.
         required: true
         type: str
         choices:
@@ -114,7 +113,7 @@ author:
 
 EXAMPLES = r"""
 ---
-- name: Manage Service Group Objects in Strata Cloud Manager
+- name: Manage Application Groups in Strata Cloud Manager
   hosts: localhost
   gather_facts: false
   vars_files:
@@ -126,88 +125,82 @@ EXAMPLES = r"""
       tsg_id: "{{ tsg_id }}"
       log_level: "INFO"
   tasks:
-    - name: Create a service group
-      cdot65.scm.service_group:
+    - name: Create web applications group
+      cdot65.scm.application_group:
         provider: "{{ provider }}"
-        name: "web-services"
+        name: "web-apps"
         members:
-          - "HTTP"
-          - "HTTPS"
-          - "SSH"
-        folder: "Texas"
-        tag:
-          - "Web"
-          - "Automation"
-        state: "present"
-
-    - name: Update service group members
-      cdot65.scm.service_group:
-        provider: "{{ provider }}"
-        name: "web-services"
-        members:
-          - "HTTP"
-          - "HTTPS"
-          - "SSH"
-          - "FTP"
+          - "ssl"
+          - "web-browsing"
         folder: "Texas"
         state: "present"
 
-    - name: Remove service group
-      cdot65.scm.service_group:
+    - name: Update web applications group membership
+      cdot65.scm.application_group:
         provider: "{{ provider }}"
-        name: "web-services"
+        name: "web-apps"
+        members:
+          - "ssl"
+          - "web-browsing"
+          - "dns-base"
+        folder: "Texas"
+        state: "present"
+
+    - name: Remove application group
+      cdot65.scm.application_group:
+        provider: "{{ provider }}"
+        name: "web-apps"
         folder: "Texas"
         state: "absent"
 """
 
 RETURN = r"""
-service_group:
-    description: Details about the service group object.
+application_group:
+    description: Details about the application group object.
     returned: when state is present
     type: dict
     sample:
         id: "123e4567-e89b-12d3-a456-426655440000"
-        name: "web-services"
+        name: "web-apps"
         members:
-          - "HTTP"
-          - "HTTPS"
-          - "SSH"
+          - "ssl"
+          - "web-browsing"
         folder: "Texas"
 """
 
 
-def build_service_group_data(module_params):
+def build_application_group_data(module_params):
     """
-    Build service group data dictionary from module parameters.
+    Build application group data dictionary from module parameters.
 
     Args:
         module_params (dict): Dictionary of module parameters
 
     Returns:
-        dict: Filtered dictionary containing only relevant service group parameters
+        dict: Filtered dictionary containing only relevant application group parameters
     """
     return {
         k: v for k, v in module_params.items() if k not in ["provider", "state"] and v is not None
     }
 
 
-def get_existing_service_group(service_group_api, service_group_data):
+def get_existing_application_group(application_group_api, application_group_data):
     """
-    Attempt to fetch an existing service group object.
+    Attempt to fetch an existing application group object.
 
     Args:
-        service_group_api: ServiceGroup API instance
-        service_group_data (dict): Service group parameters to search for
+        application_group_api: ApplicationGroup API instance
+        application_group_data (dict): Application group parameters to search for
 
     Returns:
-        tuple: (bool, object) indicating if service group exists and the service group object if found
+        tuple: (bool, object) indicating if application group exists and the group object if found
     """
     try:
-        existing = service_group_api.fetch(
-            name=service_group_data["name"],
-            folder=service_group_data.get("folder"),
-            snippet=service_group_data.get("snippet"),
-            device=service_group_data.get("device"),
+        existing = application_group_api.fetch(
+            name=application_group_data["name"],
+            folder=application_group_data.get("folder"),
+            snippet=application_group_data.get("snippet"),
+            device=application_group_data.get("device"),
         )
         return True, existing
     except NotFoundError:
@@ -216,79 +209,77 @@ def get_existing_service_group(service_group_api, service_group_data):
 
 def main():
     """
-    Main execution path for the service group object module.
+    Main execution path for the application group object module.
     """
     module = AnsibleModule(
-        argument_spec=ScmSpec.service_group_spec(),
+        argument_spec=ApplicationGroupSpec.spec(),
         supports_check_mode=True,
-        required_if=[
-            ("state", "present", ["members"]),
-        ],
+        required_if=[("state", "present", ["members"])],
     )
 
     try:
         client = get_scm_client(module)
-        service_group_api = ServiceGroup(client)
+        application_group_api = ApplicationGroup(client)
 
-        service_group_data = build_service_group_data(module.params)
-        exists, existing_service_group = get_existing_service_group(
-            service_group_api,
-            service_group_data,
+        application_group_data = build_application_group_data(module.params)
+        exists, existing_group = get_existing_application_group(
+            application_group_api,
+            application_group_data,
         )
 
         if module.params["state"] == "present":
             if not exists:
                 # Validate using Pydantic
                 try:
-                    ServiceGroupCreateModel(**service_group_data)
+                    ApplicationGroupCreateModel(**application_group_data)
                 except ValidationError as e:
                     module.fail_json(msg=str(e))
 
                 if not module.check_mode:
-                    result = service_group_api.create(data=service_group_data)
+                    result = application_group_api.create(data=application_group_data)
                     module.exit_json(
                         changed=True,
-                        service_group=serialize_response(result),
+                        application_group=serialize_response(result),
                     )
                 module.exit_json(changed=True)
             else:
                 # Compare and update if needed
                 need_update = False
-                if set(existing_service_group.members) != set(
-                    service_group_data.get("members", [])
+                if sorted(existing_group.members) != sorted(
+                        application_group_data.get("members", [])
                 ):
-                    need_update = True
-                if existing_service_group.tag != service_group_data.get("tag"):
                     need_update = True
 
                 if need_update:
                     # Prepare update data
-                    update_data = service_group_data.copy()
-                    update_data["id"] = str(existing_service_group.id)
+                    update_data = application_group_data.copy()
+                    update_data["id"] = str(existing_group.id)
 
                     # Validate using Pydantic
                     try:
-                        service_group_update_model = ServiceGroupUpdateModel(**update_data)
+                        application_group_update_model = ApplicationGroupUpdateModel(**update_data)
                     except ValidationError as e:
                         module.fail_json(msg=str(e))
 
                     if not module.check_mode:
-                        result = service_group_api.update(service_group=service_group_update_model)
+                        result = application_group_api.update(
+                            app_group=application_group_update_model
+                        )
                         module.exit_json(
                             changed=True,
-                            service_group=serialize_response(result),
+                            application_group=serialize_response(result),
                         )
                     module.exit_json(changed=True)
                 else:
                     module.exit_json(
                         changed=False,
-                        service_group=serialize_response(existing_service_group),
+                        application_group=serialize_response(existing_group),
                     )
 
         elif module.params["state"] == "absent":
             if exists:
                 if not module.check_mode:
-                    service_group_api.delete(str(existing_service_group.id))
+                    application_group_api.delete(str(existing_group.id))
                 module.exit_json(changed=True)
             module.exit_json(changed=False)
