+++ plugins/modules/tag.py
@@ -10,10 +10,10 @@
 # All rights reserved.
 
 """
-Ansible module for managing application group objects in SCM.
+Ansible module for managing tag objects in SCM.
 
-This module provides functionality to create, update, and delete application group objects
-in the SCM (Security Control Manager) system. It handles application group membership
+This module provides functionality to create, update, and delete tag objects
+in the SCM (Strata Cloud Manager) system. It handles tag attributes
 and supports check mode operations.
 """
 
@@ -23,7 +23,7 @@ __metaclass__ = type
 
 from ansible.module_utils._text import to_text
 from ansible.module_utils.basic import AnsibleModule
-from ansible_collections.cdot65.scm.plugins.module_utils.api_spec import ScmSpec  # noqa: F401
+from ansible_collections.cdot65.scm.plugins.module_utils.api_spec.tag import TagSpec  # noqa: F401
 from ansible_collections.cdot65.scm.plugins.module_utils.authenticate import (  # noqa: F401
     get_scm_client,
 )
@@ -32,47 +32,53 @@ from ansible_collections.cdot65.scm.plugins.module_utils.serialize_response impo
 )
 from pydantic import ValidationError
 
-from scm.config.objects.application_group import ApplicationGroup
+from scm.config.objects.tag import Tag
 from scm.exceptions import NotFoundError
-from scm.models.objects.application_group import (
-    ApplicationGroupCreateModel,
-    ApplicationGroupUpdateModel,
-)
+from scm.models.objects.tag import TagCreateModel, TagUpdateModel
 
 DOCUMENTATION = r"""
 ---
-module: application_group
+module: tag
 
-short_description: Manage application group objects in SCM.
+short_description: Manage tag objects in SCM.
 
 version_added: "0.1.0"
 
 description:
-    - Manage application group objects within Strata Cloud Manager (SCM).
-    - Supports creation, modification, and deletion of application groups.
-    - Ensures proper validation of group membership.
+    - Manage tag objects within Strata Cloud Manager (SCM).
+    - Supports creation, modification, and deletion of tag objects.
+    - Ensures proper validation of tag attributes including color values.
     - Ensures that exactly one of 'folder', 'snippet', or 'device' is provided.
 
 options:
     name:
-        description: The name of the application group.
+        description: The name of the tag.
         required: true
         type: str
-    members:
-        description: List of application / group / filter names.
-        required: true
-        type: list
-        elements: str
+    color:
+        description: Color associated with the tag.
+        required: false
+        type: str
+        choices:
+            - Azure Blue
+            - Black
+            - Blue
+            - Blue Gray
+            # ... (list all colors from the Colors enum)
+    comments:
+        description: Comments for the tag.
+        required: false
+        type: str
     folder:
-        description: The folder in which the resource is defined.
+        description: The folder where the tag is stored.
         required: false
         type: str
     snippet:
-        description: The snippet in which the resource is defined.
+        description: The configuration snippet for the tag.
         required: false
         type: str
     device:
-        description: The device in which the resource is defined.
+        description: The device where the tag is configured.
         required: false
         type: str
     provider:
@@ -99,7 +105,7 @@ options:
                 type: str
                 default: "INFO"
     state:
-        description: Desired state of the application group object.
+        description: Desired state of the tag object.
         required: true
         type: str
         choices:
@@ -112,7 +118,7 @@ author:
 
 EXAMPLES = r"""
 ---
-- name: Manage Application Groups in Strata Cloud Manager
+- name: Manage Tag Objects in Strata Cloud Manager
   hosts: localhost
   gather_facts: false
   vars_files:
@@ -124,82 +130,76 @@ EXAMPLES = r"""
       tsg_id: "{{ tsg_id }}"
       log_level: "INFO"
   tasks:
-    - name: Create web applications group
-      cdot65.scm.application_group:
+    - name: Create a new tag
+      cdot65.scm.tag:
         provider: "{{ provider }}"
-        name: "web-apps"
-        members:
-          - "ssl"
-          - "web-browsing"
+        name: "Production"
+        color: "Blue"
+        comments: "Production environment tag"
         folder: "Texas"
         state: "present"
 
-    - name: Update web applications group membership
-      cdot65.scm.application_group:
+    - name: Update tag color
+      cdot65.scm.tag:
         provider: "{{ provider }}"
-        name: "web-apps"
-        members:
-          - "ssl"
-          - "web-browsing"
-          - "dns-base"
+        name: "Production"
+        color: "Red"
         folder: "Texas"
         state: "present"
 
-    - name: Remove application group
-      cdot65.scm.application_group:
+    - name: Remove tag
+      cdot65.scm.tag:
         provider: "{{ provider }}"
-        name: "web-apps"
+        name: "Production"
         folder: "Texas"
         state: "absent"
 """
 
 RETURN = r"""
-application_group:
-    description: Details about the application group object.
+tag:
+    description: Details about the tag object.
     returned: when state is present
     type: dict
     sample:
         id: "123e4567-e89b-12d3-a456-426655440000"
-        name: "web-apps"
-        members:
-          - "ssl"
-          - "web-browsing"
+        name: "Production"
+        color: "Blue"
         folder: "Texas"
 """
 
 
-def build_application_group_data(module_params):
+def build_tag_data(module_params):
     """
-    Build application group data dictionary from module parameters.
+    Build tag data dictionary from module parameters.
 
     Args:
         module_params (dict): Dictionary of module parameters
 
     Returns:
-        dict: Filtered dictionary containing only relevant application group parameters
+        dict: Filtered dictionary containing only relevant tag parameters
     """
     return {
         k: v for k, v in module_params.items() if k not in ["provider", "state"] and v is not None
     }
 
 
-def get_existing_application_group(application_group_api, application_group_data):
+def get_existing_tag(tag_api, tag_data):
     """
-    Attempt to fetch an existing application group object.
+    Attempt to fetch an existing tag object.
 
     Args:
-        application_group_api: ApplicationGroup API instance
-        application_group_data (dict): Application group parameters to search for
+        tag_api: Tag API instance
+        tag_data (dict): Tag parameters to search for
 
     Returns:
-        tuple: (bool, object) indicating if application group exists and the group object if found
+        tuple: (bool, object) indicating if tag exists and the tag object if found
     """
     try:
-        existing = application_group_api.fetch(
-            name=application_group_data["name"],
-            folder=application_group_data.get("folder"),
-            snippet=application_group_data.get("snippet"),
-            device=application_group_data.get("device"),
+        existing = tag_api.fetch(
+            name=tag_data["name"],
+            folder=tag_data.get("folder"),
+            snippet=tag_data.get("snippet"),
+            device=tag_data.get("device"),
         )
         return True, existing
     except NotFoundError:
@@ -208,77 +208,74 @@ def get_existing_application_group(application_group_api, application_group_data
 
 def main():
     """
-    Main execution path for the application group object module.
+    Main execution path for the tag object module.
     """
     module = AnsibleModule(
-        argument_spec=ScmSpec.application_group_spec(),
+        argument_spec=TagSpec.spec(),
         supports_check_mode=True,
-        required_if=[("state", "present", ["members"])],
     )
 
     try:
         client = get_scm_client(module)
-        application_group_api = ApplicationGroup(client)
+        tag_api = Tag(client)
 
-        application_group_data = build_application_group_data(module.params)
-        exists, existing_group = get_existing_application_group(
-            application_group_api,
-            application_group_data,
+        tag_data = build_tag_data(module.params)
+        exists, existing_tag = get_existing_tag(
+            tag_api,
+            tag_data,
         )
 
         if module.params["state"] == "present":
             if not exists:
                 # Validate using Pydantic
                 try:
-                    ApplicationGroupCreateModel(**application_group_data)
+                    TagCreateModel(**tag_data)
                 except ValidationError as e:
                     module.fail_json(msg=str(e))
 
                 if not module.check_mode:
-                    result = application_group_api.create(data=application_group_data)
+                    result = tag_api.create(data=tag_data)
                     module.exit_json(
                         changed=True,
-                        application_group=serialize_response(result),
+                        tag=serialize_response(result),
                     )
                 module.exit_json(changed=True)
             else:
                 # Compare and update if needed
                 need_update = False
-                if sorted(existing_group.members) != sorted(
-                    application_group_data.get("members", [])
-                ):
-                    need_update = True
+                for key, value in tag_data.items():
+                    if hasattr(existing_tag, key) and getattr(existing_tag, key) != value:
+                        need_update = True
+                        break
 
                 if need_update:
                     # Prepare update data
-                    update_data = application_group_data.copy()
-                    update_data["id"] = str(existing_group.id)
+                    update_data = tag_data.copy()
+                    update_data["id"] = str(existing_tag.id)
 
                     # Validate using Pydantic
                     try:
-                        application_group_update_model = ApplicationGroupUpdateModel(**update_data)
+                        tag_update_model = TagUpdateModel(**update_data)
                     except ValidationError as e:
                         module.fail_json(msg=str(e))
 
                     if not module.check_mode:
-                        result = application_group_api.update(
-                            app_group=application_group_update_model
-                        )
+                        result = tag_api.update(tag=tag_update_model)
                         module.exit_json(
                             changed=True,
-                            application_group=serialize_response(result),
+                            tag=serialize_response(result),
                         )
                     module.exit_json(changed=True)
                 else:
                     module.exit_json(
                         changed=False,
-                        application_group=serialize_response(existing_group),
+                        tag=serialize_response(existing_tag),
                     )
 
         elif module.params["state"] == "absent":
             if exists:
                 if not module.check_mode:
-                    application_group_api.delete(str(existing_group.id))
+                    tag_api.delete(str(existing_tag.id))
                 module.exit_json(changed=True)
             module.exit_json(changed=False)
