+++ requirements.yml
@@ -1,0 +1,3 @@
+---
+roles: [ ]
+collections: [ ]
