+++ tests/ansible.cfg
@@ -1,2 +1,11 @@
 [defaults]
-vault_password_file = ./.vault_password
+collections_paths = ../collections
+stdout_callback = debug
+stderr_callback = debug
+display_skipped_hosts = yes
+
+[galaxy]
+server_list = release_galaxy
+
+[galaxy_server.release_galaxy]
+url = https://galaxy.ansible.com/
