+++ tox.ini
@@ -1,0 +1,27 @@
+[tox]
+envlist = ruff, darglint, pylint
+skipsdist = true
+requires =
+    ruff
+    darglint
+    pylint
+
+[testenv]
+allowlist_externals = *
+commands =
+    {envpython} --version
+
+[testenv:ruff]
+deps = ruff
+commands =
+    bash -c 'ruff check --exclude .tox --select ALL --ignore INP001,FA102,UP001,UP010,I001,FA100,PLR0913,E501,FBT001,C901 -q extensions/eda/plugins'
+
+[testenv:darglint]
+deps = darglint
+commands =
+    bash -c 'darglint -s numpy -z full extensions/eda/plugins'
+
+[testenv:pylint]
+deps = pylint
+commands =
+    bash -c 'find ./extensions/eda/plugins -name "*.py" -print0 | xargs -0 pylint --output-format=parseable -sn --disable R0801,E0401,C0103,R0913,R0902,R0903'
